requirements.txt
numpy==1.25.*
pillow==10.0.*
PyOpenGL==3.1.7
PyQt6==6.5.*